+++ agent/backend/src/main/java/eu/bbmri_eric/quality/agent/dataquality/dto/QualityCheckCreateDTO.java
@@ -43,7 +43,7 @@ public class QualityCheckCreateDTO {
   private Integer errorThreshold;
 
   @Schema(description = "Epsilon budget for differential privacy", example = "1.0")
-  private Float epsilonBudget;
+  private Double epsilonBudget;
 
   public QualityCheckCreateDTO() {}
 
@@ -53,7 +53,7 @@ public class QualityCheckCreateDTO {
       String query,
       Integer warningThreshold,
       Integer errorThreshold,
-      Float epsilonBudget) {
+      Double epsilonBudget) {
     this.name = name;
     this.description = description;
     this.query = query;
