+++ agent/backend/src/main/java/eu/bbmri_eric/quality/agent/dataquality/impl/InvalidConditionICDCheck.java
@@ -95,7 +95,7 @@ class InvalidConditionICDCheck implements DataQualityCheck {
   }
 
   @Override
-  public Float getEpsilonBudget() {
+  public Double getEpsilonBudget() {
     return config.getEpsilonBudget();
   }
